requirements.txt
streamlit
langchain
langchain-core
langchain-openai
langchain-community
langchain-qdrant
fastembed
pypdf
langgraph
python-dotenv